fix lexical cast error
- change npc script id type to int64 from int32

+++ sl/emulator/game/script/class/lua_npc.cpp
@@ -18,9 +18,9 @@ namespace sunlight
         );
     }
 
-    auto LuaNPC::GetId() const -> int32_t
+    auto LuaNPC::GetId() const -> uint32_t
     {
-        return static_cast<int32_t>(_npc.GetId().Unwrap());
+        return _npc.GetId().Unwrap();
     }
 
     auto LuaNPC::GetImpl() const -> GameNPC&
